+++ test/ql.jl
@@ -15,6 +15,7 @@
     )
     language = "r"
 
+    # Strict querying
     @testset "match_type=:strict, partial arguments: (NO MATCHES)" begin
         _parsed = ParSitter.parse(R_code, language)
         target = ParSitter.build_xml_tree(_parsed)
@@ -100,6 +101,7 @@
         end
     end
 
+    # Non-strict querying
     @testset "match_type=:nonstrict, partial arguments: (OK)" begin
         _parsed = ParSitter.parse(R_code, language)
         target = ParSitter.build_xml_tree(_parsed)
@@ -182,6 +184,105 @@
             target.root,
             generated_query;
             match_type = :nonstrict,
+            target_tree_nodevalue = _target_nodevalue,
+            query_tree_nodevalue = _query_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _apply_regex_glob
+        )
+        filter!(first, query_results) # keep only matches
+        @test length(query_results) == 1  # single match
+
+        CORRECT_CAPTURES = ["family" => "binomial", "id_val" => "\"linear\"", "identifier" => "link", "data" => "data_variable"]
+        _, qres = first(query_results)
+        @test length(keys(qres)) == 4
+        for (k, correct_val) in CORRECT_CAPTURES
+            @test qres[k][1].v == correct_val
+        end
+    end
+
+    # Speculative querying
+    @testset "match_type=:speculative, partial arguments: (OK)" begin
+        _parsed = ParSitter.parse(R_code, language)
+        target = ParSitter.build_xml_tree(_parsed)
+
+        query_snippet = """
+                {{comment::COMMENT}}
+                {{::IDENTIFIER}} <- glmmTMB({{::R_FORMULA}},
+                                          family ={{family::IDENTIFIER}}({{identifier::IDENTIFIER}}={{id_val::STRING}}))
+        """
+        generated_query, _, _ = ParSitter.QueryLanguage.parse_code_snippet_to_query(query_snippet, language)
+
+        query_results = ParSitter.query(
+            target.root,
+            generated_query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_nodevalue,
+            query_tree_nodevalue = _query_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _apply_regex_glob
+        )
+        filter!(first, query_results) # keep only matches
+        @test length(query_results) == 1  # single match
+
+        CORRECT_CAPTURES = [
+            "family" => "binomial",
+            "id_val" => "\"linear\"",
+            "identifier" => "link",
+            "comment" => "#acomment",
+        ]  # no spaces in comments
+        _, qres = first(query_results)
+        @test length(keys(qres)) == 4
+        for (k, correct_val) in CORRECT_CAPTURES
+            @test qres[k][1].v == correct_val
+        end
+    end
+
+    @testset "match_type=:speculative, unordered arguments: (OK)" begin
+        _parsed = ParSitter.parse(R_code, language)
+        target = ParSitter.build_xml_tree(_parsed)
+
+        query_snippet = """
+         {{::IDENTIFIER}} <- glmmTMB({{::R_FORMULA}},
+                                      family = {{family::IDENTIFIER}}({{identifier::IDENTIFIER}}={{id_val::STRING}}),
+                                      data = {{data::IDENTIFIER}})
+        """
+        generated_query, _, _ = ParSitter.QueryLanguage.parse_code_snippet_to_query(query_snippet, language)
+
+        query_results = ParSitter.query(
+            target.root,
+            generated_query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_nodevalue,
+            query_tree_nodevalue = _query_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _apply_regex_glob
+        )
+        filter!(first, query_results) # keep only matches
+        @test length(query_results) == 1  # single match
+
+        CORRECT_CAPTURES = ["family" => "binomial", "id_val" => "\"linear\"", "identifier" => "link", "data" => "data_variable"]
+        _, qres = first(query_results)
+        @test length(keys(qres)) == 4
+        for (k, correct_val) in CORRECT_CAPTURES
+            @test qres[k][1].v == correct_val
+        end
+    end
+
+    @testset "match_type=:speculative, ordered arguments: (OK)" begin
+        _parsed = ParSitter.parse(R_code, language)
+        target = ParSitter.build_xml_tree(_parsed)
+
+        query_snippet = """
+         {{::IDENTIFIER}} <- glmmTMB({{::R_FORMULA}},
+                                      data = {{data::IDENTIFIER}},
+                                      family = {{family::IDENTIFIER}}({{identifier::IDENTIFIER}}={{id_val::STRING}}))
+        """
+        generated_query, _, _ = ParSitter.QueryLanguage.parse_code_snippet_to_query(query_snippet, language)
+
+        query_results = ParSitter.query(
+            target.root,
+            generated_query;
+            match_type = :speculative,
             target_tree_nodevalue = _target_nodevalue,
             query_tree_nodevalue = _query_nodevalue,
             capture_function = _capture_function,