+++ test/query.jl
@@ -296,20 +296,348 @@ end
         @test sum(first, results) == 0  # no match
         @test sum(p -> !isempty(p[2]), results) == 2
     end
+    @testset "case 5: unordered query" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        # "1"
+        # ├─ "-3"
+        # │  └─ "*"
+        # └─ "3"
+        #    └─ "*"
+        #       └─ "@v"
+        # will match
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("1", ("-3", "*"), ("3", ("*", "@v")))
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :nonstrict,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 1  # match
+        expected_captures = [
+            MultiDict("v" => 5),
+        ]
+        for (is_match, captures) in results
+            if is_match
+                _captures = MultiDict{String, Int}(string(k) => Int.(v) for (k, v) in captures)
+                @test _captures in expected_captures
+            end
+        end
+    end
+    @testset "case 6: ambiguous query" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        # "*"
+        # ├─ "*"
+        # │  └─ "*"
+        # └─ "*"
+        #    └─ "@v"
+        # will match however match could be either v=4 or v=-4
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("*", ("*", "*"), ("*", ("@v")))
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :nonstrict,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 1  # match
+        expected_captures = [
+            MultiDict([["v" => v for v in [-4, 4]]...]),
+        ]
+        for (is_match, captures) in results
+            if is_match
+                _captures = MultiDict{String, Int}(string(k) => Int.(v) for (k, v) in captures)
+                @test _captures in expected_captures
+            end
+        end
+    end
 end
 
 
-@testset "code quering" begin
-    # TODO(Corneliu): Implement tests
-    @testset "function call" begin
-        @test true
-    end
+@testset "match_type==:speculative quering" begin
+    @testset "case 1" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        # "@v1"
+        # └─ "@v2"
+        # will match from the target:
+        # 	1->2 or 1->3 or 1->-3  # thats a single match (starting at root of tree)
+        #   3->4, -3->-4, 4->5  # three distict matches (subtrees)
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("@v1", "@v2")
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 4
+        @test sum(p -> !isempty(p[2]), results) == 4
 
-    @testset "parameters" begin
-        @test true
+        expected_captures = [
+            (MultiDict("v1" => 1, "v2" => v2) for v2 in [2, 3, -3])...,
+            MultiDict("v2" => 4, "v1" => 3),
+            MultiDict("v2" => -4, "v1" => -3),
+            MultiDict("v2" => 5, "v1" => 4),
+        ]
+        for (is_match, captures) in results
+            if is_match
+                _captures = MultiDict{String, Int}(string(k) => Int.(v) for (k, v) in captures)
+                @test _captures in expected_captures
+            end
+        end
     end
+    @testset "case 2" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        #"@v1"
+        #└─ "@v2"
+        #   └─ "@v3"
+        # will match from the target:
+        # 	1->3->4 or 1->-3->-4 # thats a single match (starting at root of tree)
+        #   3->4->5  # another tree (subtrees)
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("@v1", ("@v2", "@v3"))
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 2  # two full matches
+        @test sum(p -> !isempty(p[2]), results) == 2
 
-    @testset "edge cases..." begin
-        @test true
+        expected_captures = [
+            MultiDict("v1" => 1, "v2" => 3, "v3" => 4),
+            MultiDict("v1" => 1, "v2" => 3, "v3" => 4),
+            MultiDict("v2" => 4, "v1" => 3, "v3" => 5),
+        ]
+        for (is_match, captures) in results
+            if is_match
+                _captures = MultiDict{String, Int}(string(k) => Int.(v) for (k, v) in captures)
+                @test _captures in expected_captures
+            end
+        end
+    end
+    @testset "case 3" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        # "@v1"
+        # └─ "3"
+        #    └─ "@v3"
+        #          └─ "5"
+        # will match from the target:
+        # 	1->3->4->5
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("@v1", ("3", ("@v3", "5")))
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 1  # the only match
+        @test sum(p -> !isempty(p[2]), results) == 1
+
+        expected_captures = [
+            MultiDict("v3" => 4, "v1" => 1),
+        ]
+        for (is_match, captures) in results
+            if is_match
+                _captures = MultiDict{String, Int}(string(k) => Int.(v) for (k, v) in captures)
+                @test _captures in expected_captures
+            end
+        end
+    end
+    @testset "case 4" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        # "@v1"
+        # └─ "@v2"
+        #    ├─ "4"
+        #    └─ "5"
+        # will NOT match from the target:
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("@v1", ("@v2", "4", "5"))
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 0  # no match
+        @test sum(p -> !isempty(p[2]), results) == 2
+    end
+    @testset "case 5: unordered query" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        # "1"
+        # ├─ "-3"
+        # │  └─ "*"
+        # └─ "3"
+        #    └─ "*"
+        #       └─ "@v"
+        # will match
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("1", ("-3", "*"), ("3", ("*", "@v")))
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 1  # match
+        expected_captures = [
+            MultiDict("v" => 5),
+        ]
+        for (is_match, captures) in results
+            if is_match
+                _captures = MultiDict{String, Int}(string(k) => Int.(v) for (k, v) in captures)
+                @test _captures in expected_captures
+            end
+        end
+    end
+    @testset "case 6: ambiguous query" begin
+        # target:
+        # 1
+        # ├─ 2
+        # ├─ 3
+        # │  └─ 4
+        # │     └─ 5
+        # └─ -3
+        #    └─ -4
+        # query:
+        # "*"
+        # ├─ "*"
+        # │  └─ "*"
+        # └─ "*"
+        #    └─ "@v"
+        # will match however match could be either v=4 or v=-4
+        target = ParSitter.build_tq_tree(
+            (1, 2, (3, (4, 5)), (-3, -4))
+        )
+        query = ParSitter.build_tq_tree(
+            ("*", ("*", "*"), ("*", ("@v")))
+        )
+        results = ParSitter.query(
+            target,
+            query;
+            match_type = :speculative,
+            target_tree_nodevalue = _target_tree_nodevalue,
+            query_tree_nodevalue = _query_tree_nodevalue,
+            capture_function = _capture_function,
+            node_comparison_yields_true = _capture_on_empty_query_value
+        )
+        @test sum(first, results) == 1  # match
+        expected_captures = [
+            MultiDict("v" => -4),
+            MultiDict("v" => 4),
+        ]
+        for (is_match, captures) in results
+            if is_match
+                _captures = MultiDict{String, Int}(string(k) => Int.(v) for (k, v) in captures)
+                @test _captures in expected_captures
+            end
+        end
     end
 end