+++ NEWS.md
@@ -1,6 +1,7 @@
 v0.2.0
 ------
  - Query language support
+ - `:speculative` i.e. first match query support
  - Package tests with Aqua.jl
  - Add public API
  - Bugfixes